+++ lib/dimension.ts
@@ -1,5 +1,4 @@
-import { Entity, EntityQueryOptions, entityQueryOptionsToRaw } from "./entity.js";
-import { Location, BlockLocation } from "./location.js";
+import { Entity } from "./entity.js";
 import { map } from "./iterable.js";
 import * as MC from "@minecraft/server";
 
@@ -16,13 +15,13 @@ export class Dimension {
         return this.#dimension.id;
     }
 
-    public getEntities(opts?: EntityQueryOptions): Iterable<Entity> {
-        return map(this.#dimension.getEntities(entityQueryOptionsToRaw(opts)), raw => {
+    public getEntities(opts?: MC.EntityQueryOptions): Iterable<Entity> {
+        return map(this.#dimension.getEntities(opts), raw => {
             return new Entity(raw);
         });
     }
 
-    public spawnEntity(identifier: string, location: Location|BlockLocation): Entity {
-        return new Entity(this.#dimension.spawnEntity(identifier, location.raw));
+    public spawnEntity(identifier: string, location: MC.Vector3): Entity {
+        return new Entity(this.#dimension.spawnEntity(identifier, location));
     }
 }