+++ lib/item-stack.ts
@@ -9,21 +9,14 @@ export class ItemStack {
     public constructor(rawItemStack: MC.ItemStack);
 
     /** Construct an item stack. */
-    public constructor(itemType: MC.ItemType|string, amount?: number, data?: number);
+    public constructor(itemType: MC.ItemType|string, amount?: number);
 
     public constructor(arg0: MC.ItemStack|MC.ItemType|string, ...rest: any[]) {
         if (arg0 instanceof MC.ItemStack) {
             this.#itemStack = arg0;
         }
-        else if (arg0 instanceof MC.ItemType) {
-            this.#itemStack = new MC.ItemStack(arg0, ...rest);
-        }
         else {
-            const ty = MC.Items.get(arg0);
-            if (!ty) {
-                throw Error(`Unknown item ID: ${arg0}`);
-            }
-            this.#itemStack = new MC.ItemStack(ty, ...rest);
+            this.#itemStack = new MC.ItemStack(arg0, ...rest);
         }
     }
 
@@ -37,13 +30,6 @@ export class ItemStack {
     }
     set amount(n: number) {
         this.#itemStack.amount = n;
-    }
-
-    get data(): number {
-        return this.#itemStack.data;
-    }
-    set data(n: number) {
-        this.#itemStack.data = n;
     }
 
     get typeId(): string {